Got game state writing to work consistently
Instead of the whole server-request-client-game-state thing, I've
changed it so that the client always sends its state in every message.
This way we don't have to deal with async, and it makes more logical
sense.
I had to modify sending the chatMessage update in clientBase for this to
work. Now it calls a custom function sendChatMessage() which is defined
in client.updater.js.
Writing the game initial start state now also works as I modified the
serverBase to handle an initialization message on game start.